+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Antonio Ramirez.
 ## YiiBooster version 4.1.0
 - **(enh)** update TbHtml to the latest 1.3.0, and addeing its dependency TbArray  
 - **(fix)** TbActiveForm fix "[a]attr" format attribute #856 
+- **(fix)** allow multi relational column - #872  
 - **(enh)** 
 
 ## YiiBooster version 4.0.1